#1 refactor : 시간 표시 부분 Timeslot 컴포넌트로 분리
